% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{hello}
\alias{hello}
\title{Hello  function}
\usage{
hello(txt = "world")
}
\arguments{
\item{txt}{an object of \code{character} class}
}
\value{
\code{NULL} (invisible)
}
\description{
Hello  function
}
\details{
Here explain what function is doing and why.
}
\examples{
hello("Michal")
}
